+++ android/lib/architecture-test/src/test/java/cz/quanti/rocketapp/android/lib/architecturetest/infrastructure/KonsistExtensions.kt
@@ -1,12 +1,13 @@
 package cz.quanti.rocketapp.android.lib.architecturetest.infrastructure
 
+import com.lemonappdev.konsist.api.container.KoScope
 import com.lemonappdev.konsist.api.declaration.KoFileDeclaration
 import com.lemonappdev.konsist.api.provider.KoHasPackageProvider
 import com.lemonappdev.konsist.api.provider.KoResideInPackageProvider
-import cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model.ArchitectureLayer
-import cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model.ModuleType
-import cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model.PackagePatterns
-import cz.quanti.rocketapp.android.lib.architecturetest.infrastructure.model.Platform
+import cz.quanti.rocketapp.android.lib.architecturetest.model.ArchitectureLayer
+import cz.quanti.rocketapp.android.lib.architecturetest.model.ModuleType
+import cz.quanti.rocketapp.android.lib.architecturetest.model.PackagePatterns
+import cz.quanti.rocketapp.android.lib.architecturetest.model.Platform
 
 fun KoHasPackageProvider.resideInLayer(layer: ArchitectureLayer): Boolean {
     return hasPackage("${PackagePatterns.ROOT_PACKAGE}..${layer.layerName}..")
@@ -34,6 +35,12 @@ fun KoFileDeclaration.resideInModule(
                 "${moduleType.typeName}." +
                 "${moduleName.toModuleNameInPackage()}..",
         )
+    }
+}
+
+fun KoScope.isLayerEmpty(domain: ArchitectureLayer): Boolean {
+    return files.none { fileDeclaration ->
+        fileDeclaration.resideInLayer(domain)
     }
 }
