+++ android/lib/ui-system/src/main/java/cz/quanti/rocketapp/android/lib/uisystem/presentation/UiScreenState.kt
@@ -1,9 +1,5 @@
 package cz.quanti.rocketapp.android.lib.uisystem.presentation
 
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import cz.quanti.rocketapp.android.lib.uisystem.system.LocalSnackbar
-
 /**
  * Base class for handling with UI Screen state in cases,\
  * where there are data fetched from the server and there is \
@@ -31,39 +27,4 @@ sealed interface UiScreenState<out T> {
     data class Error(
         val errorMessage: UiText = UiText.Empty,
     ) : UiScreenState<Nothing>
-}
-
-/**
- * Show message in snackbar.
- * As per design rules, @Compose function should not result in "Nothing",
- * so the caller is responsible for checking there is a message to show.
- */
-@Composable
-fun UiScreenState.Data<*>.ShowMessageInSnackBar() {
-    val string = this.errorMessage.asString()
-    val provider = LocalSnackbar.current
-
-    LaunchedEffect(this) {
-        provider.showSnackbar(string)
-    }
-}
-
-@Composable
-fun UiScreenState.Loading.ShowMessageInSnackBar() {
-    val string = this.message.asString()
-    val provider = LocalSnackbar.current
-
-    LaunchedEffect(this) {
-        provider.showSnackbar(string)
-    }
-}
-
-@Composable
-fun UiScreenState.Error.ShowMessageInSnackBar() {
-    val string = this.errorMessage.asString()
-    val provider = LocalSnackbar.current
-
-    LaunchedEffect(this) {
-        provider.showSnackbar(string)
-    }
 }
