+++ multiplatform/feature/rocket/src/commonMain/kotlin/cz/quanti/rocketapp/multiplatform/feature/rocket/data/SpaceXApi.kt
@@ -1,5 +1,7 @@
 package cz.quanti.rocketapp.multiplatform.feature.rocket.data
 
+import cz.quanti.rocketapp.multiplatform.feature.rocket.model.RocketData
+
 internal interface SpaceXApi {
     suspend fun listRockets(): List<RocketData>
