+++ multiplatform/feature/rocket/src/commonMain/kotlin/cz/quanti/rocketapp/multiplatform/feature/rocket/data/SpaceXApiImpl.kt
@@ -1,5 +1,6 @@
 package cz.quanti.rocketapp.multiplatform.feature.rocket.data
 
+import cz.quanti.rocketapp.multiplatform.feature.rocket.model.RocketData
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
