+++ multiplatform/feature/rocket/src/commonMain/kotlin/cz/quanti/rocketapp/multiplatform/feature/rocket/domain/GetRocketsUseCaseImpl.kt
@@ -1,7 +1,7 @@
 package cz.quanti.rocketapp.multiplatform.feature.rocket.domain
 
-import cz.quanti.rocketapp.multiplatform.feature.rocket.data.RocketData
 import cz.quanti.rocketapp.multiplatform.feature.rocket.model.Rocket
+import cz.quanti.rocketapp.multiplatform.feature.rocket.model.RocketData
 import cz.quanti.rocketapp.multiplatform.lib.common.infrastructure.asResult
 import cz.quanti.rocketapp.multiplatform.lib.common.model.Result
 import kotlinx.coroutines.flow.Flow
