+++ multiplatform/feature/rocket/src/commonMain/kotlin/cz/quanti/rocketapp/multiplatform/feature/rocket/domain/RocketsRepository.kt
@@ -1,6 +1,6 @@
 package cz.quanti.rocketapp.multiplatform.feature.rocket.domain
 
-import cz.quanti.rocketapp.multiplatform.feature.rocket.data.RocketData
+import cz.quanti.rocketapp.multiplatform.feature.rocket.model.RocketData
 import kotlinx.coroutines.flow.Flow
 
 internal interface RocketsRepository {
