+++ multiplatform/feature/rocket/src/commonTest/kotlin/cz/quanti/rocketapp/multiplatform/feature/rocket/infrastructure/TestUtils.kt
@@ -1,7 +1,7 @@
 package cz.quanti.rocketapp.multiplatform.feature.rocket.infrastructure
 
 import com.goncalossilva.resources.Resource
-import cz.quanti.rocketapp.multiplatform.feature.rocket.data.RocketData
+import cz.quanti.rocketapp.multiplatform.feature.rocket.model.RocketData
 import kotlinx.serialization.json.Json
 
 private val json =
